+++ lib/sendbird_chat_sdk.dart
@@ -43,6 +43,7 @@ export 'src/public/main/model/channel/feed_channel_change_logs.dart';
 export 'src/public/main/model/channel/group_channel_change_logs.dart';
 export 'src/public/main/model/channel/group_channel_filter.dart';
 export 'src/public/main/model/channel/group_channel_unread_item_count.dart';
+export 'src/public/main/model/channel/notification_category.dart';
 export 'src/public/main/model/chat/do_not_disturb.dart';
 export 'src/public/main/model/chat/emoji.dart';
 export 'src/public/main/model/chat/global_notification_channel_setting.dart';