+++ lib/src/public/main/model/chat/snooze_period.g.dart
@@ -8,6 +8,6 @@ part of 'snooze_period.dart';
 
 SnoozePeriod _$SnoozePeriodFromJson(Map<String, dynamic> json) => SnoozePeriod(
       isSnoozeOn: json['snooze_enabled'] as bool,
-      startDate: _msToDateTime(json['snooze_start_ts'] as int),
-      endDate: _msToDateTime(json['snooze_end_ts'] as int),
+      startDate: _msToDateTime(json['snooze_start_ts']),
+      endDate: _msToDateTime(json['snooze_end_ts']),
     );